pyproject.toml
[tool.poetry]
name = "sweepy"
version = "0.1.0"
description = "Sweepstakes generator"
authors = ["Rory Devitt <rory.devitt@wegift.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.12.3"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
hypothesis = "^6.100.2"
ruff = "^0.4.2"
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
